Update CODE_OF_CONDUCT.md
I did not find this page when doing the original audit of the TOPS GitHub. Language updated in a few places to be more inclusive (adding "including but not limited to") when listing identities, and updated formatting to be consistent with other audit updates. Changes were not made to the suggested corrective actions.