M1 / L2  Quality and Diversity of Scholarly Communications.
Under the Quality and Diversity of Scholarly Communications section, remove 2nd "scientific results' from the 3rd sentence.

This results in a state where scientific results in certain disciplines published scientific results may have a number of exaggerated effects, or even be “false positives” (wrongly claiming that an effect exists), making it difficult to evaluate the trustworthiness of published results, according to a 2011 and 2016 study.