M3 / L1 FAIR: Findable, Accessible, Interoperable, Reusable
5/21: Add sentence as the 2nd sentence in the section titled FAIR: Findable, Accessible, Interoperable, Reusable "Although FAIR has been introduced in Module 2, Lesson 3, some additional details are provided below. "

 First paragraph under FAIR: Findable, Accessible, Interoperable, Reusable should read:

The vast majority of data today is shared online. Although FAIR has been introduced in Module 2, Lesson 3, some additional details are provided below. FAIR principles help researchers make better use of, and engage with a broader audience with, their scientific data than outdated techniques would allow. FAIR data are more valuable for science because they are easier to use. Data can be FAIR regardless of whether it is openly shared or not. If data are openly shared, being FAIR helps with reuse and expands the scientific impact of the data.